// src/app/models/cashier.model.ts
export interface Bill {
  id: number;
  displayOrder?: number;
  floorId?: number;
  deskId?: number;
  userCode?: string;
  userType?: string;
  customerId?: number;
  customerName?: string;
  quantityCustomer?: number;
  totalAmount?: number;
  amountReceivedByCus?: number;
  amountSendToCus?: number;
  discountPrice?: number;
  discountType?: string;
  status: string;
  note?: string;
  isPayment?: boolean;
  isPrintBill?: boolean;
  isPriority?: boolean;
  createdDate?: string;
  updatedDate?: string;
  isDeleted?: boolean;
  products?: BillDetail[];
  typePay?: string;
  totalCustomer?: number;
  debitCode?: string;
  debitDetailCodeFirst?: string;
  surcharge?: number;
  billNumber?: string;
  type?: string;
  vat?: number;
  vatCode?: string;
  vatRate?: number;
  detail1?: string;
  detail2?: string;
  billDetails?: any[];
  billDetailsRoot?: any[];
}

export interface BillDetail {
  id: number;
  billId: number;
  goodsId: number;
  quantity: number;
  unitPrice: number;
  discountPrice: number;
  discountType: string;
  createdDate?: string;
  updatedDate?: string;
  isDeleted?: boolean;
  taxVat?: number;
  goodsName?: string;
  goodsCode?: string;
  note?: string;
  image1?: string;
  discountPriceBill?: number;
  surchargeBill?: number;
}
// ...
export class NotificationCountResult {
  count: number;
}

export class NotificationResult {
  id: number;
  billId: number;
  userCode: string;
  customerName: string;
  tranType: string;
  note: string;
  status: string;
  createdDate: Date;
  displayOrder: number;
  totalAmount?: number;
  deskName?: string;
  billNumber?: string;
}